# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "Auth_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register-organizer": {
      "post": {
        "operationId": "Auth_registerOrganizer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterOrganizerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "Auth_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "Auth_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rotates the refresh token. Reusing an old one revokes every session for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "Auth_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "Auth_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "Auth_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "Auth_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired access token"
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/check-ins": {
      "post": {
        "operationId": "CheckIns_checkIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Every resolved scan is 201, including refusals like ALREADY_USED; see docs/scanner.md.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckInResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Check Ins"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** STAFF, ORGANIZER, ADMIN"
      }
    },
    "/api/v1/events/{eventId}/check-ins": {
      "get": {
        "operationId": "CheckIns_findForEvent",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Check Ins"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ORGANIZER, ADMIN"
      }
    },
...
    "/api/v1/organizer/staff": {
      "get": {
        "operationId": "EventStaff_roster",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Event Staff"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ORGANIZER"
      }
    },
    "/api/v1/events/{id}/staff": {
      "get": {
        "operationId": "EventStaff_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Event Staff"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ORGANIZER, ADMIN"
      },
      "post": {
        "operationId": "EventStaff_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Event Staff"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ORGANIZER, ADMIN"
      }
    },
    "/api/v1/events/{id}/staff/{assignmentId}": {
      "put": {
        "operationId": "EventStaff_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStaffAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Event Staff"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ORGANIZER, ADMIN"
      },
      "delete": {
        "operationId": "EventStaff_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Event Staff"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ORGANIZER, ADMIN"
      }
    },
...
  },
  "info": {
    "title": "Event Ticketing Platform API",
    "description": "REST API shared by the web app and the organizer/staff mobile apps. Amounts are integer minor units (butut, GMD). Sign in with POST /api/v1/auth/login, then send the access token as `Authorization: Bearer <token>`. See docs/api.md.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 12
          },
          "fullName": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "CUSTOMER",
          "ORGANIZER",
          "STAFF",
          "ADMIN"
        ]
      },
      "PublicUserDto": {
        "type": "object",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "role",
          "id",
          "email",
          "fullName"
        ]
      },
      "AuthSessionDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/PublicUserDto"
          },
          "accessToken": {
            "type": "string",
            "description": "Short-lived JWT (15 min by default). Send as `Authorization: Bearer <token>`."
          },
          "refreshToken": {
            "type": "string",
            "description": "Opaque, single-use. Exchange at POST /auth/refresh; store it in the Keychain/Keystore on mobile."
          }
        },
        "required": [
          "user",
          "accessToken",
          "refreshToken"
        ]
      },
      "RegisterOrganizerDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 12
          },
          "businessName": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "businessName"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "TokenPairDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string",
            "description": "Replaces the refresh token you sent, which is now revoked. Reusing the old one signs the user out everywhere."
          },
          "tokenId": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "tokenId"
        ]
      },
      "SuccessDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "minLength": 12
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "CurrentUserDto": {
        "type": "object",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "role",
          "id",
          "email"
        ]
      },
...
      "CreateCheckInDto": {
        "type": "object",
        "properties": {
          "qrToken": {
            "type": "string",
            "minLength": 32
          },
          "gateId": {
            "type": "string"
          },
          "eventId": {
            "type": "string"
          }
        },
        "required": [
          "qrToken"
        ]
      },
      "CheckInResult": {
        "type": "string",
        "enum": [
          "VALID",
          "ALREADY_USED",
          "INVALID",
          "CANCELLED",
          "REFUNDED",
          "WRONG_EVENT",
          "WRONG_DATE",
          "NO_ACCESS"
        ]
      },
      "NamedRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "TicketStatus": {
        "type": "string",
        "description": "Status before this scan (a VALID scan has just set it to USED).",
        "enum": [
          "ACTIVE",
          "USED",
          "CANCELLED",
          "REFUNDED",
          "EXPIRED",
          "TRANSFERRED"
        ]
      },
      "SeatLabelDto": {
        "type": "object",
        "properties": {
          "section": {
            "type": "string"
          },
          "row": {
            "type": "string"
          },
          "number": {
            "type": "string"
          }
        },
        "required": [
          "section",
          "row",
          "number"
        ]
      },
      "ScannedTicketDto": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "id": {
            "type": "string"
          },
          "ticketType": {
            "$ref": "#/components/schemas/NamedRefDto"
          },
          "event": {
            "description": "The ticket's own event — differs from the scanning event on WRONG_EVENT.",
            "allOf": [
              {
                "$ref": "#/components/schemas/NamedRefDto"
              }
            ]
          },
          "seat": {
            "nullable": true,
            "description": "Null for general admission.",
            "allOf": [
              {
                "$ref": "#/components/schemas/SeatLabelDto"
              }
            ]
          },
          "accessZone": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "status",
          "id",
          "ticketType",
          "event",
          "seat",
          "accessZone"
        ]
      },
      "CheckInResponseDto": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/CheckInResult"
          },
          "gate": {
            "nullable": true,
            "description": "Gate the scan was recorded at (the staff member's assigned gate if they didn't name one).",
            "allOf": [
              {
                "$ref": "#/components/schemas/NamedRefDto"
              }
            ]
          },
          "ticket": {
            "nullable": true,
            "description": "Null when the code isn't a ticket at all (INVALID).",
            "allOf": [
              {
                "$ref": "#/components/schemas/ScannedTicketDto"
              }
            ]
          }
        },
        "required": [
          "result",
          "ticket"
        ]
      },
...
      "StaffRole": {
        "type": "string",
        "enum": [
          "GATE_STAFF",
          "SCANNER_OPERATOR",
          "SECURITY",
          "MANAGER",
          "CASHIER",
          "VIP_STAFF",
          "EVENT_ADMINISTRATOR"
        ]
      },
      "AssignStaffDto": {
        "type": "object",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/StaffRole"
          },
          "email": {
            "type": "string"
          },
          "assignedGateId": {
            "type": "string"
          },
          "fullName": {
            "type": "string",
            "maxLength": 120
          },
          "password": {
            "type": "string",
            "minLength": 12
          }
        },
        "required": [
          "role",
          "email"
        ]
      },
      "UpdateStaffAssignmentDto": {
        "type": "object",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/StaffRole"
          },
          "assignedGateId": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}